+++ game_core/examples/__init__.py
@@ -10,5 +10,6 @@ from .ai_town import *
 from .ai_simulation import *
 from .spaceship import *
 from .grid import *
+from .grid_navigation import *
 from .screen_blink import *
 from .draw_system import *